+++ R/utf8-package.R
@@ -1,12 +1,3 @@
 #' @useDynLib utf8, .registration = TRUE, .fixes = "C_"
+#' @aliases NULL
 "_PACKAGE"
-
-#' @export
-encoding <- function(x) {
-  .Call(C_encoding, x)
-}
-
-#' @export
-all_utf8 <- function(x) {
-  .Call(C_all_utf8, x)
-}